refactor: extract hibernation warning message to its own component
